Revert "Add Bootstrap OSS and cloud functions in Kotlin SDK"
This reverts commit d0d6e722
